etl/requirements.txt
# ============================================================================
# AtaHub Carona - ETL Requirements
# ============================================================================

# Async HTTP Client
aiohttp==3.9.1
aiohttp[speedups]

# Async PostgreSQL Driver
asyncpg==0.29.0

# ORM and Database
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9

# Configuration Management
pydantic==2.5.2
pydantic-settings==2.1.0

# Job Scheduling
APScheduler==3.10.4

# Logging
structlog==23.2.0

# Date Utilities
python-dateutil==2.8.2

# Retry Logic (optional, built custom)
# tenacity==8.2.3

# HTTP (for compatibility)
requests==2.31.0

# Testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
